# contracts/platform-admin-openapi.yaml
openapi: 3.1.0
info:
  title: Merchant Platform Configuration Admin API
  version: 1.0.0
  description: >-
    Internal, TLS-only configuration control plane. Browser clients MUST use the
    same-origin admin BFF. Configuration values are immutable; every production
    change follows draft, approval, scheduling and fenced activation.
servers:
  - url: https://platform-admin.internal
security:
  - mutualTLS: []
    platformAssertion: []
...
paths:
...
  /internal/platform-admin/v1/providers:
    get:
      tags: [Provider operations]
      operationId: listProviderOperationBindings
      parameters: [{$ref: '#/components/parameters/Tenant'}, {name: cursor, in: query, schema: {type: string, format: uuid}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}]
      responses:
        '200': {description: Secret-free provider binding page, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderPage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/providers/{id}:
    get:
      tags: [Provider operations]
      operationId: getProviderOperationBinding
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}]
      responses:
        '200': {description: Secret-free provider binding, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderBinding'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/provider-change-requests:
    get:
      tags: [Provider operations]
      operationId: listProviderOperationChanges
      parameters: [{$ref: '#/components/parameters/Tenant'}, {name: cursor, in: query, schema: {type: string, format: uuid}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}]
      responses:
        '200': {description: Provider change page, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderChangePage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/provider-policy-requests:
    get:
      tags: [Provider operations]
      operationId: listHostedProviderPolicyRequests
      parameters: [{$ref: '#/components/parameters/Tenant'}, {name: cursor, in: query, schema: {type: string, format: uuid}}, {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}]
      responses:
        '200': {description: Secret-free hosted provider policy page, content: {application/json: {schema: {$ref: '#/components/schemas/HostedProviderPolicyPage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/providers/{id}/pause-requests:
    post:
      tags: [Provider operations]
      operationId: requestProviderPause
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderStatusRequest'}}}}
      responses:
        '202': {description: Pending provider pause, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/providers/{id}/unpause-requests:
    post:
      tags: [Provider operations]
      operationId: requestProviderUnpause
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderStatusRequest'}}}}
      responses:
        '202': {description: Pending provider unpause, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/providers/{id}/policy-requests:
    post:
      tags: [Provider operations]
      operationId: requestHostedProviderPolicy
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/HostedProviderPolicyRequest'}}}}
      responses:
        '202': {description: Pending independent provider policy approval; the bootstrap reference is never returned, content: {application/json: {schema: {$ref: '#/components/schemas/HostedProviderPolicyVersion'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/provider-change-requests/{id}/approve:
    post:
      tags: [Provider operations]
      operationId: approveProviderStatusChange
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderDecision'}}}}
      responses:
        '200': {description: Approved provider change, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /internal/platform-admin/v1/provider-change-requests/{id}/reject:
    post:
      tags: [Provider operations]
      operationId: rejectProviderStatusChange
      parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Tenant'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderDecision'}}}}
      responses:
        '200': {description: Rejected provider change, content: {application/json: {schema: {$ref: '#/components/schemas/ProviderChangeRequest'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    mutualTLS: {type: mutualTLS, description: Internal workload TLS identity enforced by the deployment boundary.}
    platformAssertion: {type: http, scheme: bearer, bearerFormat: 'PlatformAdmin compact HMAC assertion bound to method/path/query/body/audience/nonce'}
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    Tenant: {name: tenant_id, in: query, required: false, description: Omit only for a globally authorized resource, schema: {type: string, format: uuid}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 8, maxLength: 255}}
...
  responses:
    Problem: {description: Fail-closed error, content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}}
  schemas:
...
    ProviderHealthState:
      type: object
      additionalProperties: false
      required: [operation, state, error_category, version]
      properties:
        operation: {type: string, enum: [health, head, range, transaction_lookup, transfer_verify, create, status, cancel, refund, reconciliation]}
        state: {type: string, enum: [closed, open, half_open]}
        error_category: {type: string, enum: [none, timeout, dns, tls, connect, rate_limited, auth_rejected, upstream_4xx, upstream_5xx, invalid_response, chain_mismatch, genesis_mismatch, stale_head, divergent_response, policy_denied]}
        last_success_at: {type: string, format: date-time}
        last_observed_at: {type: string, format: date-time}
        lag_blocks: {type: integer, format: int64, minimum: 0}
        version: {type: integer, format: int64, minimum: 1}
    ProviderBinding:
      type: object
      additionalProperties: false
      required: [id, provider_kind, provider_id, status, version, updated_at, health]
      description: Secret-free operator projection; endpoints, credential references, account addresses, and raw errors are excluded.
      properties:
        id: {type: string, format: uuid}
        provider_kind: {type: string, enum: [on_chain, hosted]}
        provider_id: {type: string, maxLength: 128}
        tenant_id: {type: string, format: uuid}
        merchant_id: {type: string, format: uuid}
        chain_id: {type: string, maxLength: 128}
        status: {type: string, enum: [active, paused, disabled]}
        version: {type: integer, format: int64, minimum: 1}
        updated_at: {type: string, format: date-time}
        health: {type: array, maxItems: 10, items: {$ref: '#/components/schemas/ProviderHealthState'}}
    ProviderStatusRequest:
      type: object
      additionalProperties: false
      required: [expected_binding_version, reason]
      properties: {expected_binding_version: {type: integer, format: int64, minimum: 1}, reason: {type: string, minLength: 8, maxLength: 1000}}
    ProviderDecision:
      type: object
      additionalProperties: false
      required: [expected_request_version, reason]
      properties: {expected_request_version: {type: integer, format: int64, minimum: 1}, reason: {type: string, minLength: 8, maxLength: 1000}}
    ProviderChangeRequest:
      type: object
      additionalProperties: false
      required: [id, binding_id, requested_status, expected_binding_version, status, reason, requested_by, created_at, expires_at, updated_at, version]
      properties:
        id: {type: string, format: uuid}
        binding_id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        requested_status: {type: string, enum: [active, paused]}
        expected_binding_version: {type: integer, format: int64, minimum: 1}
        status: {type: string, enum: [pending_approval, completed, rejected, expired]}
        reason: {type: string, minLength: 8, maxLength: 1000}
        requested_by: {type: string, format: uuid}
        approved_by: {type: string, format: uuid}
        rejected_by: {type: string, format: uuid}
        decision_reason: {type: string, minLength: 8, maxLength: 1000}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        decided_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        version: {type: integer, format: int64, minimum: 1}
    ProviderPolicyParameters:
      type: object
      additionalProperties: false
      required: [timeout_ms, max_attempts, backoff_ms, rate_limit, rate_window_seconds, max_health_age_seconds, failure_threshold, open_seconds, half_open_successes, priority, max_lag_blocks, failure_domain]
      properties:
        timeout_ms: {type: integer, minimum: 100, maximum: 30000}
        max_attempts: {type: integer, minimum: 1, maximum: 5}
        backoff_ms: {type: integer, minimum: 0, maximum: 30000}
        rate_limit: {type: integer, minimum: 1, maximum: 10000}
        rate_window_seconds: {type: integer, minimum: 1, maximum: 3600}
        max_health_age_seconds: {type: integer, minimum: 5, maximum: 3600}
        failure_threshold: {type: integer, minimum: 1, maximum: 20}
        open_seconds: {type: integer, minimum: 1, maximum: 3600}
        half_open_successes: {type: integer, minimum: 1, maximum: 20}
        priority: {type: integer, minimum: 0, maximum: 1000}
        max_lag_blocks: {type: integer, format: int64, minimum: 0, maximum: 1000000000}
        failure_domain: {type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,127}$'}
    HostedProviderPolicies:
      type: object
      additionalProperties: false
      required: [health, create, status, cancel, refund, reconciliation]
      properties:
        health: {$ref: '#/components/schemas/ProviderPolicyParameters'}
        create: {$ref: '#/components/schemas/ProviderPolicyParameters'}
        status: {$ref: '#/components/schemas/ProviderPolicyParameters'}
        cancel: {$ref: '#/components/schemas/ProviderPolicyParameters'}
        refund: {$ref: '#/components/schemas/ProviderPolicyParameters'}
        reconciliation: {$ref: '#/components/schemas/ProviderPolicyParameters'}
    HostedProviderPolicyRequest:
      type: object
      additionalProperties: false
      required: [expected_binding_version, policies, bootstrap_probe_reference, reason]
      properties:
        expected_binding_version: {type: integer, format: int64, minimum: 1}
        policies: {$ref: '#/components/schemas/HostedProviderPolicies'}
        bootstrap_probe_reference: {type: string, minLength: 1, maxLength: 255, pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{0,254}$', writeOnly: true, description: Four-eyes bootstrap evidence input; persisted only in the immutable private policy version and never returned.}
        reason: {type: string, minLength: 8, maxLength: 1000}
    HostedProviderPolicyVersion:
      type: object
      additionalProperties: false
      required: [id, binding_id, tenant_id, policy_version, policies, payload_hash, status, expected_binding_version, reason, requested_by, created_at, expires_at, updated_at, row_version]
      description: Secret-free immutable policy projection. The approved bootstrap reference is deliberately omitted.
      properties:
        id: {type: string, format: uuid}
        binding_id: {type: string, format: uuid}
        tenant_id: {type: string, format: uuid}
        policy_version: {type: integer, format: int64, minimum: 1}
        policies: {$ref: '#/components/schemas/HostedProviderPolicies'}
        payload_hash: {type: string, pattern: '^[a-f0-9]{64}$'}
        status: {type: string, enum: [pending_approval, approved_pending_probe, active, rejected, superseded, expired]}
        expected_binding_version: {type: integer, format: int64, minimum: 1}
        reason: {type: string, minLength: 8, maxLength: 1000}
        requested_by: {type: string, format: uuid}
        approved_by: {type: string, format: uuid}
        rejected_by: {type: string, format: uuid}
        decision_reason: {type: string, minLength: 8, maxLength: 1000}
        created_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        decided_at: {type: string, format: date-time}
        activated_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        row_version: {type: integer, format: int64, minimum: 1}
...
    ProviderPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/ProviderBinding'}}, next_cursor: {type: string, format: uuid}}}
    ProviderChangePage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/ProviderChangeRequest'}}, next_cursor: {type: string, format: uuid}}}
    HostedProviderPolicyPage: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/HostedProviderPolicyVersion'}}, next_cursor: {type: string, format: uuid}}}
...
    Problem: {type: object, required: [type, title, status, detail], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, detail: {type: string}}}